requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
PyJWT==2.10.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
boto3==1.35.99
cloudinary==1.41.0
